Recorded % changes in brightnesses and smaller / less noticeable watermarking.
I tried one obvious technique where I put a line in the middle of the video, and changed it to black or white for 1s and 0s.  The message that was decoded was readable but missing a few letters.

I also split the encoded and not-encoded videos into 16x16 blocks, took the brightness, and compared them. For the most part the average brightness only changed by 1% in most areas.