Looked at excel sheets and found the macroblockers where brightness changed the least. Tried to encode into those specifically across multiple frames. Did not work well.
Made a table of all the brightnesses and hwo they changed from the native video to the values after I exported them from OpenCV - basically I wanted to know which blocks would get messed with the least if I encoded in them. I encoded in the ones that changed the least. I also commented up my code because I am starting to forget some of it.